requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.12.0
google-generativeai==0.4.0
Pillow==10.2.0
python-dotenv==1.0.0
lxml==5.1.0
numpy==1.26.4
gunicorn==21.2.0
